separate out the feed generation code
