+++ lib/account.js
@@ -521,7 +521,7 @@ export const createNote = async (body, cw, inReplyTo, toUser, editOf) => {
     const inIndex = INDEX.findIndex(idx => idx.id === object.id);
     if (inIndex < 0) {
       INDEX.push({
-        type: 'note', // someone else
+        type: 'note', // (as opposed to an activity which reps someone else)
         id: object.id,
         actor: object.attributedTo,
         published: new Date(object.published).getTime(),
