+++ lib/notes.js
@@ -1,16 +1,6 @@
 import fs from 'fs';
 import debug from 'debug';
-import { fetchUser } from './users.js';
-import {
-  addNotification,
-  getNote,
-  isReplyToMyPost,
-  isReplyToFollowing,
-  isBlocked,
-  isFollowing,
-  writeNotifications,
-  getNotifications
-} from './account.js';
+import { addNotification, isBlocked, writeNotifications, getNotifications } from './account.js';
 import {
   INDEX,
   readJSONDictionary,
@@ -160,113 +150,4 @@ const fetchActivity = async activityId => {
     });
     throw new Error('could not get post', activityId);
   }
-};
-
-export const getActivityStream = async (limit, offset) => {
-  // sort all known posts by date quickly
-  const sortedSlice = INDEX.filter(p => p.type !== 'fail' && !isNaN(p.published)).sort((a, b) => {
-    if (a.published > b.published) {
-      return -1;
-    } else if (a.published < b.published) {
-      return 1;
-    } else {
-      return 0;
-    }
-  });
-
-  const res = [];
-  let px;
-  for (px = offset; px < sortedSlice.length; px++) {
-    const p = sortedSlice[px];
-    if (p.type === 'activity') {
-      if (isFollowing(p.actor)) {
-        if (!p.inReplyTo || isReplyToMyPost(p) || (await isReplyToFollowing(p))) {
-          try {
-            const { actor } = await fetchUser(p.actor);
-            const post = await getActivity(p.id);
-            res.push({
-              note: post,
-              actor
-            });
-          } catch (err) {
-            console.error('error while loading post from index');
-          }
-        } else {
-          // disgard replies i don't care about
-        }
-      } else {
-        // disregard not from following
-      }
-    } else {
-      const post = await getNote(p.id);
-      res.push({
-        note: post,
-        actor: ActivityPub.actor
-      });
-    }
-    if (res.length === limit) {
-      break;
-    }
-  }
-
-  return {
-    activitystream: res,
-    next: px
-  };
-};
-
-export const getActivitySince = async (since, excludeSelf = false) => {
-  // sort all known posts by date quickly
-  const sortedSlice = INDEX.filter(p => p.type !== 'fail' && !isNaN(p.published))
-    .sort((a, b) => {
-      if (a.published > b.published) {
-        return -1;
-      } else if (a.published < b.published) {
-        return 1;
-      } else {
-        return 0;
-      }
-    })
-    .filter(p => {
-      if (excludeSelf && p.actor === ActivityPub.actor.id) {
-        return false;
-      }
-      return p.published > since;
-    });
-
-  const res = [];
-  let px;
-  for (px = 0; px < sortedSlice.length; px++) {
-    const p = sortedSlice[px];
-    if (p.type === 'activity') {
-      if (isFollowing(p.actor)) {
-        if (!p.inReplyTo || isReplyToMyPost(p) || (await isReplyToFollowing(p))) {
-          try {
-            const { actor } = await fetchUser(p.actor);
-            const post = await getActivity(p.id);
-            res.push({
-              note: post,
-              actor
-            });
-          } catch (err) {
-            console.error('error while loading post from index');
-          }
-        } else {
-          // disgard replies i don't care about
-        }
-      } else {
-        // disregard not from following
-      }
-    } else {
-      const post = await getNote(p.id);
-      res.push({
-        note: post,
-        actor: ActivityPub.actor
-      });
-    }
-  }
-
-  return {
-    activitystream: res
-  };
 };
